Move ServerStatus class to graylog2-plugin-interfaces.
This is needed to be able to access the server status from input
plugins.

Requires BaseConfiguration and ProcessingPauseLockedException to move as
well.
