+++ graylog2-plugin-interfaces/src/main/java/org/graylog2/plugin/ProcessingPauseLockedException.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog2.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.shared;
+package org.graylog2.plugin;
 
 /**
  * @author Lennart Koopmann <lennart@torch.sh>
