+++ graylog2-radio/src/main/java/org/graylog2/periodical/MasterPingPeriodical.java
@@ -23,7 +23,7 @@ import com.ning.http.client.AsyncHttpClient;
 import org.graylog2.plugin.periodical.Periodical;
 import org.graylog2.radio.Configuration;
 import org.graylog2.radio.cluster.Ping;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.ServerStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
