+++ graylog2-server/src/main/java/org/graylog2/Configuration.java
@@ -29,7 +29,7 @@ import com.github.joschi.jadconfig.validators.PositiveIntegerValidator;
 import com.google.common.collect.Lists;
 import com.mongodb.ServerAddress;
 import org.graylog2.plugin.Tools;
-import org.graylog2.shared.BaseConfiguration;
+import org.graylog2.plugin.BaseConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
