+++ graylog2-server/src/main/java/org/graylog2/bindings/ServerBindings.java
@@ -64,8 +64,8 @@ import org.graylog2.security.ShiroSecurityBinding;
 import org.graylog2.security.ShiroSecurityContextFactory;
 import org.graylog2.security.ldap.LdapConnector;
 import org.graylog2.security.realm.LdapUserAuthenticator;
-import org.graylog2.shared.BaseConfiguration;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.BaseConfiguration;
+import org.graylog2.plugin.ServerStatus;
 import org.graylog2.shared.bindings.AsyncHttpClientProvider;
 import org.graylog2.shared.inputs.InputRegistry;
 import org.graylog2.shared.metrics.jersey2.MetricsDynamicBinding;
