+++ graylog2-server/src/main/java/org/graylog2/bindings/providers/ServerInputRegistryProvider.java
@@ -22,7 +22,7 @@ package org.graylog2.bindings.providers;
 import org.graylog2.inputs.InputService;
 import org.graylog2.inputs.ServerInputRegistry;
 import org.graylog2.notifications.NotificationService;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.ServerStatus;
 import org.graylog2.shared.buffers.ProcessBuffer;
 import org.graylog2.shared.inputs.InputRegistry;
 import org.graylog2.shared.inputs.MessageInputFactory;
