+++ graylog2-server/src/main/java/org/graylog2/indexer/ranges/RebuildIndexRangesJob.java
@@ -28,7 +28,7 @@ import org.graylog2.indexer.Deflector;
 import org.graylog2.indexer.EmptyIndexException;
 import org.graylog2.indexer.Indexer;
 import org.graylog2.plugin.Tools;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.ServerStatus;
 import org.graylog2.system.activities.Activity;
 import org.graylog2.system.activities.ActivityWriter;
 import org.graylog2.system.jobs.SystemJob;
