+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/SystemResource.java
@@ -36,8 +36,8 @@ import org.graylog2.rest.documentation.annotations.ApiParam;
 import org.graylog2.rest.resources.RestResource;
 import org.graylog2.rest.resources.system.responses.ReaderPermissionResponse;
 import org.graylog2.security.RestPermissions;
-import org.graylog2.shared.ProcessingPauseLockedException;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.ProcessingPauseLockedException;
+import org.graylog2.plugin.ServerStatus;
 import org.graylog2.system.shutdown.GracefulShutdown;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
