+++ graylog2-server/src/main/java/org/graylog2/system/shutdown/GracefulShutdown.java
@@ -25,8 +25,8 @@ import org.graylog2.Configuration;
 import org.graylog2.initializers.BufferSynchronizerService;
 import org.graylog2.initializers.IndexerSetupService;
 import org.graylog2.plugin.lifecycles.Lifecycle;
-import org.graylog2.shared.ProcessingPauseLockedException;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.ProcessingPauseLockedException;
+import org.graylog2.plugin.ServerStatus;
 import org.graylog2.shared.initializers.InputSetupService;
 import org.graylog2.shared.initializers.PeriodicalsService;
 import org.graylog2.shared.initializers.RestApiService;
