+++ graylog2-shared/src/main/java/org/graylog2/shared/bindings/providers/EventBusProvider.java
@@ -22,7 +22,7 @@ package org.graylog2.shared.bindings.providers;
 import com.google.common.eventbus.AsyncEventBus;
 import com.google.common.eventbus.EventBus;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import org.graylog2.shared.BaseConfiguration;
+import org.graylog2.plugin.BaseConfiguration;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
