+++ graylog2-shared/src/main/java/org/graylog2/shared/bindings/providers/NodeIdProvider.java
@@ -20,7 +20,7 @@
 package org.graylog2.shared.bindings.providers;
 
 import org.graylog2.plugin.system.NodeId;
-import org.graylog2.shared.ServerStatus;
+import org.graylog2.plugin.ServerStatus;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
